fix(dashboard): wrap login search params for prerender
